+++ src/app/services/color/ldraw-color.service.ts
